+++ components/places/examples/places-utils.rs
@@ -138,6 +138,13 @@ fn run_desktop_import(db: &PlacesDb, filename: String) -> Result<()> {
     do_import(db, root)
 }
 
+fn run_ios_import(api: &PlacesApi, filename: String) -> Result<()> {
+    println!("ios import from {}", filename);
+    places::import::import_ios_bookmarks(api, filename)?;
+    println!("Import finished!");
+    Ok(())
+}
+
 fn run_native_import(db: &PlacesDb, filename: String) -> Result<()> {
     println!("import from {}", filename);
 
@@ -316,6 +323,14 @@ enum Command {
         input_file: String,
     },
 
+    #[structopt(name = "import-ios-bookmarks")]
+    /// Import bookmarks from an iOS browser.db
+    ImportIosBookmarks {
+        #[structopt(name = "input-file", long, short = "i")]
+        /// The name of the file to read.
+        input_file: String,
+    },
+
     #[structopt(name = "import-desktop-bookmarks")]
     /// Import bookmarks from JSON file exported by desktop Firefox
     ImportDesktopBookmarks {
@@ -345,6 +360,7 @@ fn main() -> Result<()> {
         } => sync(&api, engines, credential_file, wipe_all, wipe, reset),
         Command::ExportBookmarks { output_file } => run_native_export(&db, output_file),
         Command::ImportBookmarks { input_file } => run_native_import(&db, input_file),
+        Command::ImportIosBookmarks { input_file } => run_ios_import(&api, input_file),
         Command::ImportDesktopBookmarks { input_file } => run_desktop_import(&db, input_file),
     }
 }