+++ components/places/ffi/src/lib.rs
@@ -80,6 +80,18 @@ pub extern "C" fn places_connection_new(
         Ok(CONNECTIONS.insert(api.open_connection(conn_type)?))
     })
 }
+#[no_mangle]
+pub extern "C" fn places_bookmarks_import_from_ios(
+    api_handle: u64,
+    db_path: FfiStr<'_>,
+    error: &mut ExternError,
+) {
+    log::debug!("places_bookmarks_import_from_ios");
+    APIS.call_with_result(error, api_handle, |api| -> places::Result<_> {
+        places::import::import_ios_bookmarks(api, db_path.as_str())?;
+        Ok(())
+    })
+}
 
 // Best effort, ignores failure.
 #[no_mangle]