+++ CHANGELOG.md
@@ -9,6 +9,15 @@
 ### What's new
 
 - iOS only: Logins store has a new (static) `numOpenConnections` function, which can be used to detect leaks. ([#1070](https://github.com/mozilla/application-services/pull/1070))
+## Places
+
+### What's New
+
+- iOS only: PlacesApi can now migrate bookmark data from a `browser.db` database
+  via the `migrateBookmarksFromBrowserDb` function. It is recommended that this
+  only be called for non-sync users, as syncing the bookmarks over will result
+  in better handling of sync metadata, among other things.
+  ([#1078](https://github.com/mozilla/application-services/pull/1078))
 
 # v0.27.1 (_2019-04-26_)